fix: change entities, repositories and changelog, change postgresql driver to 42.4.2, applicationParams renamed

+++ src/main/java/main/repository/PersonSettingsRepository.java
@@ -1,9 +1,9 @@
 package main.repository;
 
-import main.config.entities.PersonSettings;
+import main.model.entities.PersonSettings;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PersonSettingsRepository extends JpaRepository<PersonSettings, Integer> {
+public interface PersonSettingsRepository extends JpaRepository<PersonSettings, Long> {
 }
